Split lexer whitespace handling
